spec_exec: flatten return type of run_request_speculative_fiber
Previously, it would return `Option<Result<_, _>>`. This option
was meant to represent an empty plan. This additional layer was unnecessary
and introduced additional noise.

In addition, notice that following part in speculative_execution::execute():
```
res = async_tasks.select_next_some() => {
    if let Some(r) = res {
        if !can_be_ignored(&r) {
            return r;
        } else {
            last_error = Some(r)
        }
    }
    if async_tasks.is_empty() && retries_remaining == 0 {
        return last_error.unwrap_or({
            Err(EMPTY_PLAN_ERROR)
        });
    }
}
```
would unnecessarily try to run remaining retries if `None` was returned. This means,
we would do the retries (and sleep in between retries) with empty plan.

Now, the empty plan error is returned instead of option, and can be handled
accordinly in `can_be_ignoed` (we decide not to ignore such error - no point
in retrying on another node - there is no other node).

I also added a regression test case.
The test's timeout is set to 5s. What we try to do, is to run a request
with speculative execution policy (6s retry interval) and a LBP that always
returns an empty plan. The expected behaviour is that the test completes in
less than 5s (no timeout occurs). Otherwise, this would imply that driver
was waiting 6s for speculative retry (which was the case before this commit).
I've run the test before the change, and it indeed fails.