ccm: introduce ip_allocator module
Few things happen in this commit
1. Move NetPrefix from cluster.rs to ip_allocator.rs

2. Introduce IpAllocator (with LocalSubnetIdentifier)
LocalSubnetIdentifier is a utility wrapper over two ip octets, by which
we identify a local subnet (127.x.y.0/24). It can be converted to Ipv4Addr
and vice-versa. For now, we only support ipv4 (even though NetPrefix)
supports ipv6 as well. This may change in the future.

IpAllocator does exactly what we would expect - it keeps track of the
ip address pool and allows to allocate/free addresses (subnet ids) from this pool.
Upon construction, it scans the /proc/net/tcp file to find out which subnets
are used before the tests start. This way, the pool is initialized.

We introduce a global IpAllocator for test purposes. It's fine, because
IpAllocator is thread-safe. The only assumption we make, is that the ip pool
is exclusive to the tests when they run. This is true for CI, and for most
local use cases. There are some niche cases which break this assumption,
but we do not care about it now. This solution works just fine for hackathon purposes.

I introduced two new unit tests for Ipv4Addr and LocalSubnetIdentifier conversions.

3. Adjust cluster.rs to the allocator.
The `alloc_ip_prefix` is called where NetPrefix::sniff_ip_prefix was originally called.

The `free_ip_prefix` is called in `Cluster` destructor.